+++ ajwah_block_examples/lib/todo/components/todoError.dart
@@ -3,13 +3,15 @@ import 'package:ajwah_block_examples/todo/store/TodoState.dart';
 import 'package:flutter_web/material.dart';
 
 class TodoError extends StatelessWidget {
+  final _message$ = store()
+      .select2<TodoModel>((states) => states['todo'])
+      .map((tm) => tm.message)
+      .distinct();
+
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<String>(
-      stream: store()
-          .select<TodoModel>(stateName: 'todo')
-          .map((tm) => tm.message)
-          .distinct(),
+      stream: _message$,
       builder: (BuildContext context, AsyncSnapshot<String> snapshot) {
         if (snapshot.hasData) {
           return AnimatedOpacity(